requirements.txt
aiogram
python-dotenv
pandas
openpyxl
nltk
scikit-learn